lib: move all modules files to lib directory
