+++ lib/list.js
@@ -1,5 +1,5 @@
 var fs = require('fs')
-var config = require('../config')
+var config = require('./config')
 var paths = config.paths
 
 module.exports = function () {
