+++ lib/main.js
@@ -1,7 +1,6 @@
 'use strict'
-module.exports.usage = require('./modules/usage')
-module.exports.put = require('./modules/put')
-module.exports.take = require('./modules/take')
-module.exports.list = require('./modules/list')
-module.exports.open = require('./modules/open')
-module.exports.close = require('./modules/close')
+module.exports.put = require('./put')
+module.exports.take = require('./take')
+module.exports.list = require('./list')
+module.exports.open = require('./open')
+module.exports.close = require('./close')
