+++ lib/put.js
@@ -1,7 +1,7 @@
 var fs = require('fs')
 var diff = require('diff')
 var inquirer = require('inquirer')
-var config = require('../config')
+var config = require('./config')
 var paths = config.paths
 
 module.exports = function (files, silent) {
