+++ lib/take.js
@@ -1,6 +1,6 @@
 var fs = require('fs')
 var diff = require('diff')
-var config = require('../config')
+var config = require('./config')
 var paths = config.paths
 
 module.exports = function (files) {
